.github/skills/powerpoint/requirements.txt
python-pptx
cairosvg
Pillow>=9.0.0
